Bend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
pycryptodome==3.20.0
python-dotenv==1.0.1
Werkzeug==3.0.1